# openapi.json
{
    "openapi":  "3.0.1",
    "info":  {
                 "title":  "OpenAPI definition",
                 "version":  "v0"
             },
    "servers":  [
                    {
                        "url":  "http://localhost:8080",
                        "description":  "Generated server url"
                    }
                ],
    "security":  [
                     {
                         "bearerAuth":  [

                                        ]
                     }
                 ],
...
    "paths":  {
                  "/api/supplier/orders/{id}/status":  {
                                                           "put":  {
                                                                       "tags":  [
                                                                                    "supplier-controller"
                                                                                ],
                                                                       "operationId":  "updateOrderStatus",
                                                                       "parameters":  [
                                                                                          {
                                                                                              "name":  "id",
                                                                                              "in":  "path",
                                                                                              "required":  true,
                                                                                              "schema":  "@{type=integer; format=int64}"
                                                                                          },
                                                                                          {
                                                                                              "name":  "status",
                                                                                              "in":  "query",
                                                                                              "required":  true,
                                                                                              "schema":  "@{type=string}"
                                                                                          }
                                                                                      ],
                                                                       "responses":  {
                                                                                         "200":  {
                                                                                                     "description":  "OK",
                                                                                                     "content":  "@{*/*=}"
                                                                                                 }
                                                                                     }
                                                                   }
                                                       },
                  "/api/supplier/inventory/{id}":  {
                                                       "put":  {
                                                                   "tags":  [
                                                                                "supplier-controller"
                                                                            ],
                                                                   "operationId":  "updateInventory",
                                                                   "parameters":  [
                                                                                      {
                                                                                          "name":  "id",
                                                                                          "in":  "path",
                                                                                          "required":  true,
                                                                                          "schema":  "@{type=integer; format=int64}"
                                                                                      }
                                                                                  ],
                                                                   "requestBody":  {
                                                                                       "content":  {
                                                                                                       "application/json":  "@{schema=}"
                                                                                                   },
                                                                                       "required":  true
                                                                                   },
                                                                   "responses":  {
                                                                                     "200":  {
                                                                                                 "description":  "OK",
                                                                                                 "content":  "@{*/*=}"
                                                                                             }
                                                                                 }
                                                               },
                                                       "delete":  {
                                                                      "tags":  [
                                                                                   "supplier-controller"
                                                                               ],
                                                                      "operationId":  "deleteInventory",
                                                                      "parameters":  [
                                                                                         {
                                                                                             "name":  "id",
                                                                                             "in":  "path",
                                                                                             "required":  true,
                                                                                             "schema":  "@{type=integer; format=int64}"
                                                                                         }
                                                                                     ],
                                                                      "responses":  {
                                                                                        "200":  {
                                                                                                    "description":  "OK",
                                                                                                    "content":  "@{*/*=}"
                                                                                                }
                                                                                    }
                                                                  }
                                                   },
                  "/api/retailer/inventory/{id}":  {
                                                       "put":  {
                                                                   "tags":  [
                                                                                "retailer-controller"
                                                                            ],
                                                                   "operationId":  "updateInventory_1",
                                                                   "parameters":  [
                                                                                      {
                                                                                          "name":  "id",
                                                                                          "in":  "path",
                                                                                          "required":  true,
                                                                                          "schema":  "@{type=integer; format=int64}"
                                                                                      }
                                                                                  ],
                                                                   "requestBody":  {
                                                                                       "content":  {
                                                                                                       "application/json":  "@{schema=}"
                                                                                                   },
                                                                                       "required":  true
                                                                                   },
                                                                   "responses":  {
                                                                                     "200":  {
                                                                                                 "description":  "OK",
                                                                                                 "content":  "@{*/*=}"
                                                                                             }
                                                                                 }
                                                               },
                                                       "delete":  {
                                                                      "tags":  [
                                                                                   "retailer-controller"
                                                                               ],
                                                                      "operationId":  "deleteInventory_1",
                                                                      "parameters":  [
                                                                                         {
                                                                                             "name":  "id",
                                                                                             "in":  "path",
                                                                                             "required":  true,
                                                                                             "schema":  "@{type=integer; format=int64}"
                                                                                         }
                                                                                     ],
                                                                      "responses":  {
                                                                                        "200":  {
                                                                                                    "description":  "OK",
                                                                                                    "content":  "@{*/*=}"
                                                                                                }
                                                                                    }
                                                                  }
                                                   },
...
                  "/api/supplier/inventory":  {
                                                  "get":  {
                                                              "tags":  [
                                                                           "supplier-controller"
                                                                       ],
                                                              "operationId":  "getInventory",
                                                              "responses":  {
                                                                                "200":  {
                                                                                            "description":  "OK",
                                                                                            "content":  "@{*/*=}"
                                                                                        }
                                                                            }
                                                          },
                                                  "post":  {
                                                               "tags":  [
                                                                            "supplier-controller"
                                                                        ],
                                                               "operationId":  "addInventory",
                                                               "requestBody":  {
                                                                                   "content":  {
                                                                                                   "application/json":  "@{schema=}"
                                                                                               },
                                                                                   "required":  true
                                                                               },
                                                               "responses":  {
                                                                                 "200":  {
                                                                                             "description":  "OK",
                                                                                             "content":  "@{*/*=}"
                                                                                         }
                                                                             }
                                                           }
                                              },
                  "/api/retailer/inventory":  {
                                                  "get":  {
                                                              "tags":  [
                                                                           "retailer-controller"
                                                                       ],
                                                              "operationId":  "getInventory_1",
                                                              "responses":  {
                                                                                "200":  {
                                                                                            "description":  "OK",
                                                                                            "content":  "@{*/*=}"
                                                                                        }
                                                                            }
                                                          },
                                                  "post":  {
                                                               "tags":  [
                                                                            "retailer-controller"
                                                                        ],
                                                               "operationId":  "addInventory_1",
                                                               "requestBody":  {
                                                                                   "content":  {
                                                                                                   "application/json":  "@{schema=}"
                                                                                               },
                                                                                   "required":  true
                                                                               },
                                                               "responses":  {
                                                                                 "200":  {
                                                                                             "description":  "OK",
                                                                                             "content":  "@{*/*=}"
                                                                                         }
                                                                             }
                                                           }
                                              },
...
              },
    "components":  {
...
                       "securitySchemes":  {
                                               "bearerAuth":  {
                                                                  "type":  "http",
                                                                  "scheme":  "bearer",
                                                                  "bearerFormat":  "JWT"
                                                              }
                                           }
                   }
}